Enable strict C++ compiler warnings with -Werror
Enable a comprehensive set of compiler warnings for all CUTLASS targets
(but not third-party code like Google Test) and treat them as errors so
no new warnings can be introduced going forward.

Warnings enabled: -Wall -Wextra -Wshadow -Wnon-virtual-dtor -Wcast-align
-Woverloaded-virtual -Wmisleading-indentation -Wnull-dereference -Wformat=2
-Wimplicit-fallthrough -Wduplicated-cond -Wlogical-op (GCC-only) -Werror.

Warnings intentionally suppressed for template-heavy CUDA code:
-Wno-unused-parameter -Wno-unused-local-typedefs -Wno-unknown-pragmas
-Wno-class-memaccess -Wno-comment -Wno-deprecated-declarations.

Key changes:
- CMakeLists.txt: Add warning flags applied per-target via
  cutlass_apply_standard_compile_options() so third-party code is not
  affected. Windows builds upgraded from /W3 to /W4 /WX.
- ~100 library headers: Fix constructor parameter shadowing member
  variables by renaming parameters with trailing underscores.
- CuTe headers: Fix shadow warnings in layout.hpp, stride.hpp, and
  various arch/atom/util headers.
- Numeric type headers: Fix sign-conversion in half.h, shadow in
  float8.h, array-bounds suppression in array_subbyte.h.
- Test files: Fix shadow, unused-but-set, and attribute warnings
  across all unit test suites (core, cute, gemm, conv, epilogue,
  transform, reduction).

+++ include/cutlass/numeric_conversion.h
@@ -2328,7 +2328,7 @@ struct NumericArrayConverterPacked4Element<float_ue8m0_t, float, FloatRoundStyle
     NumericConverter<result_element, source_element, FloatRoundStyle::round_toward_infinity> converter;
 
     CUTLASS_PRAGMA_UNROLL
-    for (int i = 0; i < 4; ++i) {
+    for (size_t i = 0; i < 4; ++i) {
       result[i] = converter(source[i]);
     }
     return result;
@@ -2372,7 +2372,7 @@ struct NumericArrayConverterPacked4Element<float_ue8m0_t, float, FloatRoundStyle
     NumericConverter<result_element, source_element, FloatRoundStyle::round_toward_zero> converter;
 
     CUTLASS_PRAGMA_UNROLL
-    for (int i = 0; i < 4; ++i) {
+    for (size_t i = 0; i < 4; ++i) {
       result[i] = converter(source[i]);
     }
 
@@ -3414,7 +3414,7 @@ struct NumericArrayConverter<float_ue8m0_t, float, 2, FloatRoundStyle::round_tow
     NumericConverter<result_element, source_element, FloatRoundStyle::round_toward_infinity> converter;
 
     CUTLASS_PRAGMA_UNROLL
-    for (int i = 0; i < 2; ++i) {
+    for (size_t i = 0; i < 2; ++i) {
       result[i] = converter(source[i]);
     }
 
@@ -3455,7 +3455,7 @@ struct NumericArrayConverter<float_ue8m0_t, float, 2, FloatRoundStyle::round_tow
     NumericConverter<result_element, source_element, FloatRoundStyle::round_toward_zero> converter;
 
     CUTLASS_PRAGMA_UNROLL
-    for (int i = 0; i < 2; ++i) {
+    for (size_t i = 0; i < 2; ++i) {
       result[i] = converter(source[i]);
     }
 
@@ -5645,12 +5645,12 @@ struct NumericArrayConverter<int8_t, int4b_t, N, Round> {
     NumericConverter<int8_t, int4b_t, Round> convert_;
     
     CUTLASS_PRAGMA_UNROLL
-    for (int i = 0; i < N; ++i) {
+    for (size_t i = 0; i < static_cast<size_t>(N); ++i) {
       result[i] = convert_(source[i]);
     }
-    
+
     return result;
-    
+
     #endif // __CUDA_ARCH__
   }
